+++ src/app/chat/ChatBot/style.ts
@@ -4,8 +4,6 @@ const useStyles = createStyles(({ css }) => ({
   chatBot: css`
     display: flex;
     flex-direction: column;
-    flex-grow: 1;
-
     width: 100%;
     height: 100%;
   `,
