+++ src/features/ChatInfo/index.tsx
@@ -3,11 +3,9 @@
 import { DraggablePanel } from '@lobehub/ui';
 import { Button } from 'antd';
 import { createStyles } from 'antd-style';
-import { useState } from 'react';
 
 import AgentInfo from '@/components/AgentInfo';
 import Voice from '@/features/ChatInput/Actions/Voice';
-import { agentListSelectors, useAgentStore } from '@/store/agent';
 import { useConfigStore } from '@/store/config';
 import { sessionSelectors, useSessionStore } from '@/store/session';
 
@@ -24,12 +22,6 @@ const useStyles = createStyles(({ css, token }) => ({
 
 const Header = () => {
   const { styles } = useStyles();
-  const [tempId, setTempId] = useState<string>('');
-  const [showAgentSidebar, activateAgent, deactivateAgent] = useAgentStore((s) => [
-    agentListSelectors.showSideBar(s),
-    s.activateAgent,
-    s.deactivateAgent,
-  ]);
   const [openPanel] = useConfigStore((s) => [s.openPanel, s.closePanel]);
   const [currentAgent] = useSessionStore((s) => [sessionSelectors.currentAgent(s)]);
 
@@ -37,18 +29,9 @@ const Header = () => {
     <DraggablePanel
       classNames={{ content: styles.content }}
       defaultSize={{ width: 360 }}
-      expand={showAgentSidebar}
       minWidth={280}
       maxWidth={420}
       mode={'fixed'}
-      onExpandChange={(show) => {
-        if (!show) {
-          setTempId(useAgentStore.getState().currentIdentifier);
-          deactivateAgent();
-        } else if (tempId) {
-          activateAgent(tempId);
-        }
-      }}
       placement={'right'}
     >
       <AgentInfo
