+++ src/layout/AppLayout.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { createStyles } from 'antd-style';
 import { ReactNode, memo } from 'react';
+import { Flexbox } from 'react-layout-kit';
 
 import Apps from '@/layout/Apps';
 import Background from '@/layout/Background';
@@ -14,24 +14,14 @@ interface AppLayoutDesktopProps {
   headerKey?: HeaderNavKey;
 }
 
-const useStyles = createStyles(({ css }) => ({
-  content: css`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: calc(100vh - 64px);
-  `,
-}));
-
 const AppLayoutDesktop = memo<AppLayoutDesktopProps>(({ children, headerKey }) => {
-  const { styles } = useStyles();
   return (
-    <>
+    <Flexbox height={'100%'} width={'100%'}>
       <Header headerKey={headerKey} />
-      <div className={styles.content}>{children}</div>
+      {children}
       <Apps />
       <Background />
-    </>
+    </Flexbox>
   );
 });
