+++ src/panels/AgentPanel/Market/index.tsx
@@ -1,6 +1,6 @@
 import { createStyles } from 'antd-style';
-import classNames from 'classnames';
 import React from 'react';
+import { Flexbox } from 'react-layout-kit';
 
 import TopBanner from '@/components/TopBanner';
 
@@ -8,15 +8,6 @@ import AgentCard from './Card';
 import AgentList from './List';
 
 const useStyles = createStyles(({ css }) => ({
-  market: css`
-    position: relative;
-
-    display: flex;
-
-    width: 100%;
-    height: 100%;
-    min-height: 500px;
-  `,
   container: css`
     overflow-y: auto;
 
@@ -27,7 +18,6 @@ const useStyles = createStyles(({ css }) => ({
   `,
   content: css`
     max-width: 1024px;
-    height: 100%;
     margin: 0 auto;
   `,
   title: css`
@@ -38,17 +28,11 @@ const useStyles = createStyles(({ css }) => ({
   `,
 }));
 
-interface AgentProps {
-  className?: string;
-  style?: React.CSSProperties;
-}
-
-const Agent = (props: AgentProps) => {
+const Agent = () => {
   const { styles } = useStyles();
-  const { style, className } = props;
 
   return (
-    <div className={classNames(className, styles.market)} style={style}>
+    <Flexbox flex={1} height={'100%'} width={'100%'} horizontal>
       <div className={styles.container}>
         <div className={styles.content}>
           <TopBanner title="Find Your Lovest Vidol" />
@@ -56,7 +40,7 @@ const Agent = (props: AgentProps) => {
         </div>
       </div>
       <AgentCard />
-    </div>
+    </Flexbox>
   );
 };
